refactor(internal): retire feature bridge residue
