+++ packages/internal/README.md
@@ -18,6 +18,7 @@ Do not install this package directly. Other ViteHub packages depend on it inside
 // packages/*/src/internal-use.ts
 import { createRuntimeRegistryContents } from "@vite-hub/internal/definition-catalog"
 import { createNoExternalMerger } from "@vite-hub/internal/build/vite"
+import { normalizeHosting } from "@vite-hub/internal/hosting"
 ```
 
 ## Used by
