+++ packages/internal/package.json
@@ -13,7 +13,7 @@
     "./arrays": "./dist/arrays.js",
     "./definition-catalog": "./dist/definition-catalog.js",
     "./definition-discovery": "./dist/definition-discovery.js",
-    "./feature-bridge/hosting": "./dist/feature-bridge/hosting.js",
+    "./hosting": "./dist/hosting.js",
     "./integrations/hex": "./dist/integrations/hex.js",
     "./object": "./dist/object.js",
     "./provision": "./dist/provision.js",
