+++ packages/internal/src/hosting.ts
