+++ packages/internal/vite.config.ts
@@ -25,7 +25,7 @@ export default defineConfig({
       "src/build/vercel-config.ts",
       "src/build/vercel-runtime-packages.ts",
       "src/build/vite.ts",
-      "src/feature-bridge/hosting.ts",
+      "src/hosting.ts",
       "src/integrations/hex.ts",
       "src/runtime/app.ts",
       "src/runtime/cloudflare-env.ts",
